trading-bot/requirements.txt
# Core trading framework
freqtrade==2023.12
ccxt==4.2.25

# Data analysis and manipulation
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.11.0

# Machine learning and AI
tensorflow>=2.16.0
torch>=2.2.0
scikit-learn>=1.4.0
transformers==4.36.2
huggingface-hub==0.19.4

# Technical analysis
TA-Lib==0.4.28
talib-binary==0.4.19

# Data sources
yfinance==0.2.28
alpha-vantage==2.3.1
requests==2.31.0
websocket-client==1.6.4

# Database and API
psycopg2-binary==2.9.9
redis==5.0.1
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.2

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
loguru==0.7.2
colorama==0.4.6
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development tools
black==23.12.1
flake8==6.1.0
mypy==1.8.0
jupyter==1.0.0
ipykernel==6.29.0

# Additional AI libraries
openai==1.6.1
anthropic==0.7.8
langchain==0.1.0
chromadb==0.4.22

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Cryptocurrency specific
cryptofeed==2.4.0
pycoingecko==3.1.0
